change staffid to use long

+++ cs3343.group18.usermanagementsystem/src/usermanagementsystem/datastructure_interface/IUserInfo.java
@@ -12,7 +12,7 @@ public interface IUserInfo {
 
     EnumGender getGender();
 
-    int getStaffId();
+    long getStaffId();
 
     EnumPosition getPosition();
